ДЗ_5.1 разделение на модули
