+++ .gitignore
@@ -11,3 +11,5 @@ npm-debug.log*
 Thumbs.db
 .vscode/
 get_repeats.js
+test.js
+main_dz4.js
