+++ index.html
@@ -234,9 +234,10 @@
       </section>
     </template>
 
-    <script src="./js/main.js"></script>
-    <script src="./js/functions.js"></script>
+    <script src="./js/main.js" type="module"></script>
+    <!-- <script src="./js/functions.js"></script> -->
     <!-- <script src="./js/test.js"></script> -->
+    <!-- <script src="./js/get_repeats.js"></script> -->
 
 </body>
 </html>
