+++ js/data.js
@@ -1,0 +1,81 @@
+
+
+const DESCRIPTIONS = [
+  'Курорт',
+  'указатель пути',
+  'Красивый пляж',
+  'Сиськи и фотик',
+  'типа супчик',
+  'машина Бетмена',
+  'урожай клубники не удался',
+  'А ты налей и отойди',
+  'Спасение с необитаемого острова',
+  'Кто первый того и тапки',
+  'вход в лабиринт',
+  'красивая белая машинка',
+  'тошниловка',
+  'котамбургер',
+  'в детстве мало пароли',
+  'пролетая над Парижем',
+  'хор мальчиков-зайчиков',
+  'я в машинах не разбираюст, но эта мне нравится',
+  'хочу такие тапочки с подсветкой',
+  'пальмы',
+  'а нормальной еды нет?',
+  'ухожу в закат',
+  'обитатель моря',
+  'слушают руками',
+  'Вам помочь?'
+];
+
+const COMMENTS_ARR = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Услада',
+  'Полина-Полина',
+  'Голуба',
+  'Апрель',
+  'Вишня',
+  'Принцесса Даниэлла',
+  'Заря-Заряница',
+  'Луна',
+  'Луналика',
+  'Ярослав',
+  'Лютобор',
+  'Каспер',
+  'Ненаглядный',
+  'Дельфин'
+];
+
+const PHOTOS_COUNT_MAX = 25;
+const LIKES = {
+  MIN: 15,
+  MAX: 200
+};
+
+const COMMENT = {
+  MIN: 1,
+  MAX: 30
+};
+
+const AVA = {
+  MIN: 1,
+  MAX: 6
+};
+
+export {
+  DESCRIPTIONS,
+  COMMENTS_ARR,
+  NAMES,
+  PHOTOS_COUNT_MAX,
+  LIKES,
+  COMMENT,
+  AVA
+};
